// src/client/toteco/models/establishment-dto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * TOTECO SERVICE
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: 1.0.0
 * Contact: ire.cunba@gmail.com
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

 /**
 * 
 *
 * @export
 * @interface EstablishmentDTO
 */
export interface EstablishmentDTO {

    /**
     * @type {string}
     * @memberof EstablishmentDTO
     */
    name?: string;

    /**
     * @type {string}
     * @memberof EstablishmentDTO
     */
    location?: string;

    /**
     * @type {boolean}
     * @memberof EstablishmentDTO
     */
    isComputerAllowed?: boolean;

    /**
     * @type {string}
     * @memberof EstablishmentDTO
     */
    mapsId?: string;
}